+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 - New create-, update-, detail-, and deleteviews for PokedexEntry objects.
 
   - The detailview for a PokedexEntry is the same as for species, but with slightly more information.
+
+- Added "owner"-field to Pokedex.
+
+- Added "description"-field to PokedexEntry.
+
+  - This field will override the species' description field in the detail view.